+++ Assets/_Scripts/BuildableController.cs
@@ -1,0 +1,10 @@
+using UnityEngine;
+using UnityEngine.UI;
+
+namespace _Scripts
+{
+    public class MenuItemController : MonoBehaviour
+    {
+
+    }
+}