+++ Assets/_Scripts/Player/PlayerCameraController.cs
@@ -6,8 +6,8 @@ namespace _Scripts.Player
     {
         // the first person camera
         public GameObject camera;
-        public float mouseSensitivity = 100f;
-        private float xRotation;
+        [SerializeField] private float mouseSensitivity = 100f;
+        private float _xRotation;
         
         void Start()
         {
@@ -21,11 +21,11 @@ namespace _Scripts.Player
             var mouseX = Input.GetAxis("Mouse X") * mouseSensitivity * Time.deltaTime;
             var mouseY = Input.GetAxis("Mouse Y") * mouseSensitivity * Time.deltaTime;
 
-            xRotation -= mouseY;
-            xRotation = Mathf.Clamp(xRotation, -70f, 60f);
+            _xRotation -= mouseY;
+            _xRotation = Mathf.Clamp(_xRotation, -70f, 60f);
             
             transform.Rotate(Vector3.up * mouseX);
-            camera.transform.localRotation = Quaternion.Euler(xRotation, 0f, 0f);
+            camera.transform.localRotation = Quaternion.Euler(_xRotation, 0f, 0f);
         }
     }
 }